pyproject.toml
# In pyproject.toml

[project]
name = "materials-transformer"
version = "0.1.0"
description = ""
authors = [
    {name = "hexa",email = "ethanmick741@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch (>=2.7.1,<3.0.0)",
    "torchvision (>=0.22.1,<0.23.0)",
    "torchmetrics (>=1.7.2,<2.0.0)",
    "pytorch-lightning (>=2.5.1.post0,<3.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "rich (>=14.0.0,<15.0.0)",
    "transformers (>=4.52.4,<5.0.0)",
    "datasets",
    "wandb",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "scikit-image (>=0.25.2,<0.26.0)",
    "kfp>=2.13.0",
    "mlflow>=3.1.0",
    "optuna>=4.3.0",
    "pyrootutils>=1.0.4",
    "ruff>=0.11.13",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128" },
]
torchvision = [
    { index = "pytorch-cu128" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
